Написанный код разделен на модули main.js, data.js, util.js. Созданы файлы map.js, adform.js для будущего кодирования
