+++ js/data.js
@@ -1,0 +1,15 @@
+// Вводные данные
+
+const TITLES = ['Роскошный дворец на берегу моря', 'Уютная квартирка на крыше самого высокого небоскреба', 'Затерянный домик в горах', 'Бунгало в непролазных джунглях'];
+const TYPES = ['palace','flat','house','bungalow'];
+const CHECKINS = ['12:00','13:00','14:00'];
+const CHECKOUTS = ['12:00','13:00','14:00'];
+const FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner']
+const DESCRIPTIONS = ['Чистенько и уютненько', 'Вызывающая роскошь', 'Убежище для техногика', 'Логово супергероя', 'Скромно и опрятно', 'Жилище для пилотов в душе', 'Ночной кошмар для аккуратиста'];
+const PHOTOS = [
+  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel3.jpg'];
+const NUMBER_OF_ADVERTS = 10;
+
+export {TITLES, TYPES, CHECKINS, CHECKOUTS, FEATURES, DESCRIPTIONS, PHOTOS, NUMBER_OF_ADVERTS}
