+++ js/map.js
@@ -1,0 +1,1 @@
+//работа с данными карты
